+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * fix empty lines in prompt (visible when using typing animation) [#734](https://github.com/jcubic/jquery.terminal/issues/734)
 * fix trimming whitespace with echo keep_words [#738](https://github.com/jcubic/jquery.terminal/issues/738)
 * fix JSON working on processing nested formatting with html entities [#735](https://github.com/jcubic/jquery.terminal/issues/735)
+* fix echo array of numbers [#747](https://github.com/jcubic/jquery.terminal/issues/747)
 
 ## 2.31.1
 ### Bugfix